+++ lib/airbrake-ruby/performance_breakdown.rb
@@ -12,6 +12,7 @@ module Airbrake
     include HashKeyable
     include Ignorable
     include Stashable
+    include Mergeable
 
     def initialize(
       method:,