+++ lib/airbrake-ruby/request.rb
@@ -9,6 +9,7 @@ module Airbrake
     include HashKeyable
     include Ignorable
     include Stashable
+    include Mergeable
 
     def initialize(
       method:,